requirements.txt
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
pandas==2.3.1
numpy==2.3.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3
python-jose[cryptography]==3.3.0